Change printf conversion specifiers for size_t arg
Some printf-style format strings use a "%d" signed integer conversion
specifier while casting a size_t arg (from sizeof) to an integer.
This commit changes those to a "%lu" unsigned long integer conversion
specifier while casting the sizeof value to an unsigned long.

The "%zu" size_t conversion specifier is not being used since it
does not appear until the 1999 C Standard.  Builds currently use the
default "gnu90" language standard (the GNU dialect of ISO C90 with
some C99 features).

Reference:
- http://www.embedded.com/electronics-blogs/programming-pointers/4026076/Why-size-t-matters
- http://www.embedded.com/electronics-blogs/programming-pointers/4026118/Further-insights-into-size-t
- http://www.embedded.com/electronics-blogs/programming-pointers/4007254/A-case-study-in-portability

+++ server-test.c
@@ -107,8 +107,9 @@ int parse_test_opts(
     }
     if (strlcpy(buf, str, sizeof(buf)) >= sizeof(buf)) {
         if ((errbuf != NULL) && (errlen > 0)) {
-            snprintf(errbuf, errlen, "testopts string exceeds %d-byte maximum",
-                (int) sizeof(buf) - 1);
+            snprintf(errbuf, errlen,
+                "testopts string exceeds %lu-byte maximum",
+                (unsigned long) sizeof(buf) - 1);
         }
         return(-1);
     }
